refactor: 전반적인 간단한 리팩터링
